requirements.txt
# Core Data Science
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.10.0

# External Boosting Models i.e not with scikit_learn
xgboost>=1.7.0
lightgbm>=3.3.0
catboost>=1.1.0


# Note: ydata-profiling removed as it's not compatible with Python 3.13
# and not used in the Streamlit app

# Utilities
pyyaml>=6.0

# Pin SHAP version for Python 3.13 compatibility
shap>=0.48.0,<0.49.0

# model exporting - pin version for compatibility
dill>=0.3.6

# API
flask

# install it as package on system
-e .